# resources/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "臺中市政府秘書處資訊管理系統 API",
    "version": "1.0.0",
    "description": "這是臺中市政府秘書處集合各項自行開發系統之API資源，除政府開放資料資源無須授權即可取用外，其餘資源如有需要介接使用，建請洽本處聯絡人林管理師。",
    "contact": {
      "name": "Lin, Ying-Chin",
      "email": "ar0660@taichung.gov.tw"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/mit-license.php"
    }
  },
  "host": "api.secret.taichung.gov.tw",
  "basePath": "/v1.0",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "取得所有使用者資料",
        "operationId": "UserController_getUsers",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "新增使用者",
        "operationId": "UserController_register",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "帳號",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "密碼",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "姓名",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/{uid}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "取得特定使用者資料",
        "operationId": "UserController_getUsers_uid",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "使用者編號",
            "required": true,
            "type": "number"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "刪除特定使用者",
        "operationId": "UserController_delete",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "使用者編號",
            "required": true,
            "type": "number"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/login": {
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "使用者登入",
        "operationId": "UserController_login",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "帳號",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "密碼",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "info": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "data": {
          "type": "object"
        },
        "time": {
          "type": "number"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "number"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "Token": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "token_type": {
          "type": "string",
          "example": "Bearer"
        },
        "expires_in": {
          "type": "number"
        }
      }
    },
...
  }
}